+++ CHANGELOG.md
@@ -1,26 +1,40 @@
 # PDF Generator for Craft 3
 
+## 0.1.0 - 2022-11-12
+
+- Feature update for Craft CMS 3
+- Add author option
+- Add keywords option
+- Enable manual page break mode (no_auto_page_break) option
+- Add option to password protect of PDF
+
 ## 0.0.9 - 2021-11-11 [CRITICAL]
+
 - Fix of fontDir bug, thanks to @iwi-hi
 
 ## 0.0.8 - 2021-11-10
-- merged with @AramLoosman
-- remove unused function
-- add variables to footer and header
-- changed fontDirs in plugin code to fontDir as in README.md
+
+- Merged with @AramLoosman
+- Remove unused function
+- Add variables to footer and header
+- Changed fontDirs in plugin code to fontDir as in README.md
 
 ## 0.0.7 - 2022-02-01
+
 - Add custom fonts to include in template
 
 ## 0.0.6 - 2021-10-14
+
 - Added custom title to attributes
 - Added custom variables to attributes
 - Update README.MD
 
 ## 0.0.5 - 2021-06-21
+
 - Updated README.MD
 - Rename the plugin
 
 ## 0.0.2 - 2021-03-21
+
 - Initial release
 - PDF Generator for Craft 3